Former-commit-id: 7baf0af1db79045b84e7811acb982b7d405e95a5
Former-commit-id: d6a2fd64824ef722152f44d84c61c4b76c111372

+++ App/event.vert
@@ -1,4 +1,5 @@
-#version 130
+#version 110
+#extension GL_EXT_gpu_shader4 : enable
 /**
  * @brief Source code of the vertex shader used for drawing of single-channel events.
  *
@@ -18,9 +19,9 @@ uniform int divideBy;
 
 void main()
 {
-	float x = bufferOffset + gl_VertexID/divideBy;
+	float x = float(bufferOffset) + float(gl_VertexID)/float(divideBy);
 
-	float y = y0 + yScale*sampleValue + eventWidth*(1 - 2*(gl_VertexID & 1));
+	float y = y0 + yScale*sampleValue + eventWidth*float(1 - 2*(gl_VertexID & 1));
 
 	gl_Position = transformMatrix*vec4(x, y, 0, 1);
 }
